Move Option to a separate file
