+++ src/implementations/react/src/playground/sections/DropdownSection.js
@@ -3,6 +3,17 @@ import PlaygroundSection from '../PlaygroundSection';
 import { Dropdown } from '../../hig-react';
 
 class DropdownSection extends Component {
+  constructor(props) {
+    super(props)
+    this.state = {
+      value: 'bar value'
+    }
+  }
+
+  setValue = (value) => {
+    this.setState({ value });
+  }
+
   render() {
     return (
       <PlaygroundSection title="Dropdown">
@@ -11,18 +22,18 @@ class DropdownSection extends Component {
           instructions="instructions for regular Controlled dropdown"
           placeholder="placeholder for regular Controlled dropdown"
           options={
-            [ 
-              { 
-                label: "foo", 
-                value: "foo value" 
-              }, { 
-                label: "bar", 
-                value: "bar value" 
+            [
+              {
+                label: "foo",
+                value: "foo value"
+              }, {
+                label: "bar",
+                value: "bar value"
               }
             ]
           }
-          value="bar value"
-          onChange={(id) => { console.log(id); }}
+          value={this.state.value}
+          onChange={this.setValue}
         />
 
         <Dropdown
@@ -30,13 +41,13 @@ class DropdownSection extends Component {
           instructions="instructions for regular Uncontrolled dropdown"
           placeholder="placeholder for regular Uncontrolled dropdown"
           options={
-            [ 
-              { 
-                label: "foo", 
-                value: "foo value" 
-              }, { 
-                label: "bar", 
-                value: "bar value" 
+            [
+              {
+                label: "foo",
+                value: "foo value"
+              }, {
+                label: "bar",
+                value: "bar value"
               }
             ]
           }
